requirements.txt
discord.py==1.7.3
requests==2.22.0
discord==1.7.3
python-dotenv==0.19.0
